initial js refactor
